Reorganize span/buffer helpers into Touki.Buffers and Touki.Io namespaces
Move RunLengthEncoder and BufferScope to Touki.Buffers; move SpanReader, SpanWriter, and SpanReaderExtensions to Touki.Io. Wire up global usings and update test collateral, fuzz targets, and docs accordingly.
