+++ touki.tests/Docs/SampleTests.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
-using Touki.Io;
 using Touki.Text;
 
 namespace Touki.Docs;
