+++ touki.tests/Touki/Io/SpanReaderEdgeCaseTests.cs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
-namespace Touki;
+namespace Touki.Io;
 
 /// <summary>
 /// Edge case and bug validation tests for SpanReader{T}.
