+++ touki/Framework/Polyfills/System.Globalization/DateTimeFormat.cs
@@ -5,7 +5,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Touki;
 using Touki.Framework.Resources;
 
 namespace System.Globalization;
