+++ touki/Touki/Io/SpanWriter.cs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 // See LICENSE file in the project root for full license information
 
-namespace Touki;
+namespace Touki.Io;
 
 /// <summary>
 ///  Fast stack based <see cref="Span{T}"/> writer.
